added first draft of Entrance.moveToDiningRoom().
This is the method used to move students from entrance to the dining room. In my idea it will work together with a doActions() method which should come from the PlayerController or from a dedicate actionPhaseController. At the moment this method is only some commented pseudo-code in the Player class