added first draft of Entrance.moveToIsland().
This is the other main method for the action phase; it lets a player move a selected number of students to the islands. then the player can select the island by choosing the index of the island in the map. In the future this choice will come from the gui and will be handled by a controller.