added GameMap.checkJoins() and GameMap.doJoins().
These 2 methods work together, and the 2nd should be called every time mother nature lands on an island, but only if this causes the owner to change. If that island cannot be joined, checkJoins() returns "none". otherwise it returns the position of the islands it has to be joined with.